pyproject.toml
[tool.poetry]
name = "ai-traject"
version = "0.1.0"
description = ""
authors = ["Liam <liam.goethals@inthepocket.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
langchain = "^0.2.0"
openai = "^1.30.1"
pathlib = "^1.0.1"
langchain-community = "^0.2.0"
langchain-openai = "^0.1.7"
azure-search-documents = "^11.4.0"
azure-identity = "^1.16.0"
pypdf = "^4.2.0"
requests = "^2.32.1"
thefuzz = "^0.22.1"
markdown-it-py = "^3.0.0"
markdown-to-json = "^2.1.1"
icecream = "^2.1.3"
pytest = "^8.2.1"
pytest-cov = "^5.0.0"
extract-msg = "^0.48.5"
pandas = "^2.2.2"
numpy = "^1.26.4"
ipykernel = "^6.29.4"
streamlit = "^1.34.0"
fastapi = "^0.111.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
